[mqbus-send] Properly detect running clients
There was a bug where we would remove the listener if a new listener
advertising the same id was detected. However, removing the listener
caused us to unlink the named pipe. Also existing clients would also be
detected twice: first to discover the pipe via the filesystem then
through the message queue so we'd always remove the pipe and thus
prevent future mqbus-send instances from detecting the waiting
receivers.