Sanely handle messages greater than MAX_MSG_SIZE
Previously this program silently truncated large messages which was
about the worst thing it could do. Now we'll forward the entire message
if reading from stdin. In order to have some since of atomicity, if
all of what is available form stdin won't fit in MAX_MSG_SIZE, it will
be split up be '\n'. If there are no such character is present, then the
message will just be broken up into MAX_MSG_SIZE chunks and sent
individually.

How the message is split only matters if there is risk of another client
also trying to concurrently send messages and if that's a problem.